Moved logs to cpp files rather than headers
I moved the declaration of the loggers to the cpp files rather than the
header files. This is because in order to prevent variable collisions,
sometimes very complicated names had to be used. Also, it is more logical
and appropriate to have variables, no matter their type, be declared in
the dynamic part, rather than the common/header.