Removed Wt, and started work on WebInterface
Wt is licensed under GPL, and this project is intended to be licensed under LGPL.
Those two aren't compatible, as using Wt would require Firestarter to be licensed under the GPL as well.

For this reason, I've decided to not use Wt, and go with fastcgi++ and ctemplates instead. I'm sure it's possible
to achieve some functionality along the lines of Wt, without the added cost.